change the definition of getBaseUrl() to return the appName's root URL
Since this is shared between Survey and Tables, returning the URL to the system directory (the location of the Survey index.html) does not make sense. Change this to return the root URL for the appName. E.g., http://localhost:8635/default/  This API is only used within Survey and only impacts the core logic.