simplify webView state and interactions
(1) Remove CommonApplication.configureView(); this allows us to remove getWebKitResourceId(). The CommonApplication.triggerDatabaseEvent() calls and the onPause() event delegation in the fragments containing the webkit will now properly handle updating the state of the webkit. 
(2) drop doAction notifications on the floor until the webkit has loaded. This is the new model for doAction notifications. odkData response notifications are always transmitted to the webkit. This is a change/correction for ODK Survey, but is how it was expected to work within ODK Tables.